CPP using strings
My first program writes a function 'ReformatDate' which will take the input String and returns the new format of the string. It uses for loops to edit the format by using subtraction from the old date to create snippets of the new date and align them together in the main menthod.