The clock time loops
My second program finds the maximum valid time that is displayed by the clock in a 24 hour format. The method is Maxtime which takes the four digits as the inputs(bool) and uses for loops and if statements to check for the most suitable time. It returns “not possible” if the time is not valid.
For eg: the numbers 1 9 3 4 will give the time- 19:43